nozzle: some updates
